requirements.txt
asyncpg==0.29.0
fastapi==0.111.1
pydantic==2.8.2
uvicorn==0.30.3
httpx==0.27.0
pytest==8.3.1
python-dotenv==1.0.1